Fix: Fixes issue with performance and animation
Remove inefficent code that was constantly updating the decks. This was cause major performance issues and was causing other aspects of the interface not too work.

Removal of the code broken the Deck List view from updating. So a work around using props had to be put in place. This is not an ideal situation, but it works for now.