Feat: Returns back to the swipe right to show delete
This was removed and replaced with edit button, because it did not work on iPhone. However, the problem with this feature was solved with the last commit.

Although the edit button provide the potential for multiple delete, the swipe right work better for the current interface. The code was removed to keep the code clean.