Fix: Fixes a bug with create a new card
Several problems arose when the Add Card button moved to the status bar. Several work around were required to properly pass props to the Add Card component and successfully update the Deck View.

One work around which needs to be looked at is the use of a setTimeout to delay the return back to the Deck page until the RefreshDeck function has finished executing.