feat: Adds default decks for first time users
When a user first loads the app, default decks will saved to there local storage and displayed. These decks can be deleted and will not return. A users object is now created for the app is launch for the first time. It is not use for anything right now, but will be useful in the future.